Handle the case label column already exists and forceIndexLabel = true.
